Made use of map[string]string for nw options in default-gateway impl
